+++ frontend/src/app/state/app.reducer.ts
@@ -47,7 +47,7 @@ export type AppAction =
     | { type: 'app:discardPlaylist' }
     | { type: 'app:updatePlaylist'; playlist: Playlist }
     | { type: 'app:savePlaylist' }
-    | { type: 'app:removePlaylist'; playlist: Playlist }
+    | { type: 'app:removePlaylist'; playlist: Playlist, keep: boolean }
     | { type: 'app:loadPlaylistInfo'; playlistInfo: PlaylistInfo }
     | { type: 'app:confirmDeletePlaylist'; confirmationProps: ConfirmationProps };